Compute species impact indicator

Compute species impact indicator

Usage

species_impact(
  cube,
  impact_data = NULL,
  col_category = NULL,
  col_species = NULL,
  col_mechanism = NULL,
  trans = 1,
  type = NULL
)

Arguments

cube

The data cube of class sim_cube or processed_cube from b3gbi::process_cube()

impact_data

The dataframe of species impact which contains columns of impact_category, scientific_name and impact_mechanism

col_category

The name of the column containing the impact categories. The first two letters each categories must be an EICAT short names (e.g "MC - Minimal concern")

col_species

The name of the column containing species names

col_mechanism

The name of the column containing mechanisms of impact

trans

Numeric. The type of transformation to convert the EICAT categories to numerical values. 1 converts ("MC", "MN", "MO", "MR", "MV") to (0,1,2,3,4) 2 converts ("MC", "MN", "MO", "MR", "MV") to (1,2,3,4,5) and 3 converts ("MC", "MN", "MO", "MR", "MV") to (1,10,100,1000,10000)

type

The type indicators based on the aggregation of within and across species in a site. The type can be precautionary, precautionary cumulative, mean, mean cumulative or cumulative.

Value

A dataframe of impact indicator per species (class species_impact)

Examples

acacia_cube <- taxa_cube(
  taxa = taxa_Acacia,
  region = southAfrica_sf,
  res = 0.25,
  first_year = 2010
)

speciesImpact <- species_impact(
  cube = acacia_cube,
  impact_data = eicat_acacia,
  trans = 1,
  type = "mean"
)