GBIF occurrences data of acacia in South Africa An example of occurrence data from GBIF containing required column for impact indicator.

GBIF occurrences data of acacia in South Africa An example of occurrence data from GBIF containing required column for impact indicator.

Usage

taxa_Acacia

Format

A dataframe object containing 17,279 rows and 6 variables

decimalLatitude

geographic latitude in decimal

decimalLongitude

geographic longitude in decimal

species

scientific name of species

speciesKey

GBIF species identification number

coordinateUncertaintyInMeters

radius of the uncertainty circle around geographic point

year

year occurrence was recorded

Source

https://www.gbif.org/occurrence/search?country=ZA&taxon_key=2978223&occurrence_status=present

Examples

  head(taxa_Acacia,10)
#>    decimalLatitude decimalLongitude          species speciesKey
#> 1        -34.47472         19.90503   Acacia saligna    2978552
#> 2        -34.05180         23.37747   Acacia cyclops    2980425
#> 3        -33.29982         26.50064  Acacia mearnsii    2979775
#> 4        -33.30021         26.50089 Acacia baileyana    2981303
#> 5        -33.30114         26.50165 Acacia baileyana    2981303
#> 6        -33.30096         26.50253  Acacia mearnsii    2979775
#> 7        -34.34901         19.09172   Acacia saligna    2978552
#> 8        -34.35345         19.08895   Acacia cyclops    2980425
#> 9        -34.35366         19.08878   Acacia cyclops    2980425
#> 10       -33.99522         23.52011  Acacia mearnsii    2979775
#>    coordinateUncertaintyInMeters year
#> 1                             25 2024
#> 2                             NA 2024
#> 3                              5 2024
#> 4                              5 2024
#> 5                              5 2024
#> 6                              5 2024
#> 7                              4 2024
#> 8                             14 2024
#> 9                             10 2024
#> 10                             4 2024